cookbook/examples/apps/github_repo_analyzer/requirements.txt
agno>=0.1.0
PyGithub>=2.1.1
python-dotenv>=1.0.0
matplotlib>=3.7.2
pandas>=2.0.3
streamlit>=1.24.0 
openai>=1.67.0
